+++ src/utils.js
@@ -1,4 +1,8 @@
 const ESC_KEYCODE = 27;
+const RenderPosition = {
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`
+};
 
 const getRndArrFromArr = (array, length = array.length) => {
   if (length > array.length) {
@@ -24,10 +28,28 @@ const createElm = (template) => {
   return newElm.firstChild;
 };
 
+const renderElm = (container, elm, place = `beforeend`) => {
+  switch (place) {
+    case RenderPosition.AFTERBEGIN:
+      container.prepend(elm);
+      break;
+    case RenderPosition.BEFOREEND:
+      container.append(elm);
+      break;
+  }
+};
+
 const checkKeyCode = (cb, evt) => {
   if (evt.keyCode === ESC_KEYCODE) {
     cb(evt);
   }
 };
 
-export {getRndArrFromArr, getRandomIntegerNumber, createElm, checkKeyCode};
+export {
+  RenderPosition,
+  getRndArrFromArr,
+  getRandomIntegerNumber,
+  createElm,
+  renderElm,
+  checkKeyCode,
+};
